Allow targeting css to siblings of inpage footnotes
`-cr-only-if: following-inpage-footnote` is true for a given node if the
node is a sibling of an inpage footnote, but is not itself an inpage
footnote.

In other words, it is true for (the chain of) siblings between / after
inpage footnotes, up to the end of the parent node.

This allows writing a generic style hint that extends inpage footnotes
regardless of how the nodes are marked as footnotes

With this one can write e.g.

    *, autoBoxing {
        -cr-only-if: following-inpage-footnote -inpage-footnote;
        -cr-hint: extend-footnote-inpage;
    }

or

    *:not(h1, h2, h3, h4, h5, h6), autoBoxing {
        -cr-only-if: following-inpage-footnote -inpage-footnote;
        -cr-hint: extend-footnote-inpage;
    }

To extend inpage footnotes according to the given selector, regardless of
the specific criteria by which footnotes are selected.

Instead of having to write this for every single footnote class:

    .footnote ~ *, .footnote ~ autoBoxing, ... { ... }